cleanup analyze env modules
The call_command and is_localhost functions
were used only in one module no sepatare
env modules is required for that.
